Move IssueTracked in separate module
It needs because new mixin class for isssue tracker
integration will be implemented. And it makes more
sense to contain those mixins in separate module.
